restaurant management system/requirements.txt

# =====================================================
# WEB FRAMEWORK
# =====================================================
Flask==2.3.3
Werkzeug==2.3.7

# =====================================================
# DATABASE & ORM
# =====================================================
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23

# =====================================================
# MYSQL DATABASE DRIVERS
# =====================================================
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# =====================================================
# AUTHENTICATION & SECURITY
# =====================================================
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.0.1
PyJWT==2.8.0
cryptography==41.0.7

# =====================================================
# FORM VALIDATION
# =====================================================
email-validator==2.0.0
WTForms[email]==3.0.1

# =====================================================
# ENVIRONMENT & CONFIGURATION
# =====================================================
python-dotenv==1.0.0

# =====================================================
# TEMPLATE ENGINE
# =====================================================
Jinja2==3.1.6
MarkupSafe==3.0.2

# =====================================================
# HTTP & UTILITIES
# =====================================================
requests==2.31.0
urllib3==2.1.0
click==8.2.1
itsdangerous==2.2.0
blinker==1.9.0

# =====================================================
# DATE & TIME HANDLING
# =====================================================
python-dateutil==2.8.2

# =====================================================
# DEVELOPMENT & TESTING (Optional)
# =====================================================
# Uncomment the following for development:
# pytest==7.4.3
# pytest-flask==1.3.0
# Flask-Testing==0.8.1

# =====================================================
# PRODUCTION SERVER (Optional)
# =====================================================
# Uncomment for production deployment:
# gunicorn==21.2.0
# uWSGI==2.0.23
